refactor(core): extract consolidator.ts from operations.ts
Move consolidate() and compressObservations() into dedicated
consolidator.ts module. Re-export consolidate from operations.ts
to maintain backward compatibility with all transports.
